requirements.txt
amqp==5.0.6
arabic-reshaper==2.1.3
asgiref==3.3.4
billiard==3.6.4.0
Brotli==1.0.9
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.0
cffi==1.14.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.4.1
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-celery-beat==2.2.0
django-easy-pdf==0.1.1
django-easy-pdf3==0.1.2
django-heroku==0.3.1
django-redis==5.0.0
django-timezone-field==4.1.2
django-widget-tweaks==1.4.8
django-widgets==0.1.16
dpd-components==0.1.0
Flask==2.0.1
Flask-Compress==1.9.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
Pillow==8.2.0
plotly==4.14.3
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
reportlab==3.5.67
retrying==1.3.3
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
